Move Document Upload container to a common location
