+++ submit-web/src/components/SubmissionItem/ConsultationRecord/DocumentUploadSection.tsx
@@ -3,7 +3,7 @@ import { Box, Divider, Grid, Typography } from "@mui/material";
 import { BCDesignTokens, EAOColors } from "epic.theme";
 import { useDocumentUploadStore } from "@/store/documentUploadStore";
 import DocumentContainer from "../ManagementPlanSubmission/DocumentContainer";
-import DocumentToUploadContainer from "../ManagementPlanSubmission/DocumentToUploadContainer";
+import DocumentToUploadContainer from "../DocumentToUploadContainer";
 import { When } from "react-if";
 import { Navigate, useParams } from "@tanstack/react-router";
 import { notify } from "@/components/Shared/Snackbar/snackbarStore";
