+++ submit-web/src/components/SubmissionItem/DocumentToUploadContainer.tsx
@@ -1,13 +1,13 @@
 import { Box, Grid, Typography } from "@mui/material";
 import { BCDesignTokens } from "epic.theme";
 import React, { useEffect } from "react";
-import DocumentIcon from "./DocumentIcon";
+import DocumentIcon from "./ManagementPlanSubmission/DocumentIcon";
 import { notify } from "@/components/Shared/Snackbar/snackbarStore";
 import { saveObject } from "@/hooks/api/useObjectStorage";
 import { createSubmission } from "@/hooks/api/useSubmissions";
 import { Document, useDocumentUploadStore } from "@/store/documentUploadStore";
 import { useParams } from "@tanstack/react-router";
-import ProgressBar from "./ProgressBar";
+import ProgressBar from "./ManagementPlanSubmission/ProgressBar";
 import { useQueryClient } from "@tanstack/react-query";
 import { SUBMISSION_TYPE } from "@/models/Submission";
